:wrench: 关系计算V1.0

+++ src/main/java/com/nuc/tracking/teacherend/po/record/StudentCourse.kt
@@ -16,5 +16,5 @@ class StudentCourse {
     var courseId: Long = 0
     var studentId: Long = 0
     var percent: Float = 0f
-    var tqPercent: Long = 0
+    var tqPercent: Float = 0f
 }
